refactor(cache): extract options and trade cache to mixins (#510)
Split sqlite_cache.py into domain-specific modules:
- src/cache/options_cache.py: OptionsCacheMixin (329 lines)
- src/cache/trade_cache.py: TradeCacheMixin (263 lines)

TradingCacheManager now inherits from both mixins.
Reduces sqlite_cache.py from 1587 to 947 lines (40% reduction).
