+++ src/cache/sqlite_cache.py
@@ -3,9 +3,10 @@ SQLite-based cache for market data.
 
 Replaces file-based caching with efficient relational storage.
 Provides drop-in replacement for UnifiedCacheManager with better performance.
+
+Issue #510: Split into mixins for options and trade cache functionality.
 """
 
-import json
 import logging
 import sqlite3
 import threading
@@ -16,12 +17,14 @@ from typing import Any, Dict, List, Optional
 
 import pandas as pd
 
+from src.cache.options_cache import OptionsCacheMixin
+from src.cache.trade_cache import TradeCacheMixin
 from src.utils.date_utils import get_datetime_now, now_iso, parse_date_string
 
 logger = logging.getLogger(__name__)
 
 
-class TradingCacheManager:
+class TradingCacheManager(OptionsCacheMixin, TradeCacheMixin):
     """
     SQLite-based market data cache.
 
@@ -859,650 +862,7 @@ class TradingCacheManager:
         except Exception as e:
             self.logger.error(f"Error vacuuming database: {e}")
 
-    # === RAW OPTIONS CHAIN METHODS (Issue #373) ===
-
-    def store_raw_options(
-        self,
-        symbol: str,
-        trading_date: str,
-        options_df: pd.DataFrame,
-        underlying_price: float = None,
-        source: str = "polygon",
-        data_quality_score: float = 1.0,
-        provider_metadata: dict = None,
-    ) -> int:
-        """
-        Store raw options chain data to database with multi-provider support.
-
-        Args:
-            symbol: Stock symbol (e.g., "SPY")
-            trading_date: Trading date in YYYY-MM-DD format
-            options_df: DataFrame with raw options data
-            underlying_price: Spot price of underlying (optional)
-            source: Data source ("polygon", "alpha_vantage", "alpaca", etc.)
-            data_quality_score: Data quality score 0.0-1.0 (default: 1.0)
-            provider_metadata: Optional dict with provider-specific metadata
-
-        Returns:
-            Number of options contracts stored
-
-        Expected DataFrame columns:
-            - strike (required)
-            - option_type or type (required): 'call' or 'put'
-            - expiration (required)
-            - bid, ask, last, volume, open_interest (optional)
-            - implied_volatility, delta, gamma, theta, vega, rho (optional)
-            - contract_symbol (optional)
-
-        Example:
-            >>> options_df = pd.DataFrame({
-            ...     'strike': [400, 405, 410],
-            ...     'type': ['call', 'call', 'put'],
-            ...     'expiration': ['2024-12-20', '2024-12-20', '2024-12-20'],
-            ...     'last': [10.5, 8.2, 3.1]
-            ... })
-            >>> metadata = {'api_version': '2.0', 'rate_limit': 5}
-            >>> count = cache.store_raw_options("SPY", "2024-01-15", options_df,
-            ...                                  underlying_price=412.50,
-            ...                                  source="polygon",
-            ...                                  data_quality_score=0.95,
-            ...                                  provider_metadata=metadata)
-        """
-        if options_df is None or options_df.empty:
-            self.logger.warning(f"Empty options data for {symbol} {trading_date}")
-            return 0
-
-        try:
-            # Prepare records for batch insert
-            records = []
-            df = options_df.copy()
-
-            # Normalize option_type column (handle both 'type' and 'option_type')
-            if "type" in df.columns and "option_type" not in df.columns:
-                df["option_type"] = df["type"]
-            elif "option_type" not in df.columns:
-                raise ValueError("Options data must have 'type' or 'option_type' column")
-
-            # Normalize option_type values to lowercase
-            df["option_type"] = df["option_type"].str.lower()
-
-            for _, row in df.iterrows():
-                try:
-                    # Handle expiration date conversion
-                    expiration = row["expiration"]
-                    if hasattr(expiration, "strftime"):
-                        expiration_str = expiration.strftime("%Y-%m-%d")
-                    else:
-                        expiration_str = str(expiration)
-
-                    # Handle contract_symbol (may be missing)
-                    contract_sym = row.get("contract_symbol") or row.get("contractID")
-
-                    # Use provided underlying_price or try to get from row
-                    underlying = underlying_price
-                    if underlying is None and "underlying_price" in row:
-                        underlying = row["underlying_price"]
-
-                    # Serialize provider metadata to JSON if provided
-                    metadata_json = json.dumps(provider_metadata) if provider_metadata else None
-
-                    current_time = now_iso()
-
-                    record = (
-                        symbol.upper(),
-                        trading_date,
-                        float(row["strike"]),
-                        str(row["option_type"]),
-                        expiration_str,
-                        float(row["bid"]) if pd.notna(row.get("bid")) else None,
-                        float(row["ask"]) if pd.notna(row.get("ask")) else None,
-                        float(row["last"]) if pd.notna(row.get("last")) else None,
-                        int(row["volume"]) if pd.notna(row.get("volume")) else None,
-                        int(row["open_interest"]) if pd.notna(row.get("open_interest")) else None,
-                        (
-                            float(row["implied_volatility"])
-                            if pd.notna(row.get("implied_volatility"))
-                            else None
-                        ),
-                        float(row["delta"]) if pd.notna(row.get("delta")) else None,
-                        float(row["gamma"]) if pd.notna(row.get("gamma")) else None,
-                        float(row["theta"]) if pd.notna(row.get("theta")) else None,
-                        float(row["vega"]) if pd.notna(row.get("vega")) else None,
-                        float(row["rho"]) if pd.notna(row.get("rho")) else None,
-                        contract_sym,
-                        float(underlying) if underlying is not None else None,
-                        source,
-                        current_time,  # cached_at
-                        current_time,  # modified_at
-                        float(data_quality_score),  # data_quality_score
-                        metadata_json,  # provider_metadata
-                    )
-                    records.append(record)
-
-                except Exception as e:
-                    self.logger.warning(f"Error preparing option record: {e}")
-                    continue
-
-            if not records:
-                self.logger.warning(f"No valid records prepared for {symbol} {trading_date}")
-                return 0
-
-            # Bulk insert with write lock
-            with self._write_lock:
-                with sqlite3.connect(self.db_path) as conn:
-                    cursor = conn.cursor()
-                    cursor.executemany(
-                        """
-                        INSERT OR REPLACE INTO raw_options_chain
-                        (symbol, trading_date, strike, option_type, expiration,
-                         bid, ask, last, volume, open_interest,
-                         implied_volatility, delta, gamma, theta, vega, rho,
-                         contract_symbol, underlying_price, source, cached_at,
-                         modified_at, data_quality_score, provider_metadata)
-                        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                    """,
-                        records,
-                    )
-                    conn.commit()
-                    rows_inserted = cursor.rowcount
-
-            self.logger.info(
-                f"Stored {rows_inserted} raw options for {symbol} {trading_date} "
-                f"from {source} (quality: {data_quality_score:.2f})"
-            )
-            return rows_inserted
-
-        except Exception as e:
-            self.logger.error(
-                f"Error storing raw options for {symbol} {trading_date}: {e}", exc_info=True
-            )
-            return 0
-
-    def get_raw_options(
-        self, symbol: str, trading_date: str, source: str = None
-    ) -> Optional[pd.DataFrame]:
-        """
-        Retrieve raw options chain from database with file cache fallback.
-
-        This implements the database-first architecture from Issue #373.
-        Falls back to file cache if database is empty (for backward compatibility).
-
-        Args:
-            symbol: Stock symbol (e.g., "SPY")
-            trading_date: Trading date in YYYY-MM-DD format
-            source: Data source filter (None = any source)
-
-        Returns:
-            DataFrame with raw options data, or None if not found
-
-        Example:
-            >>> options = cache.get_raw_options("SPY", "2024-01-15")
-            >>> if options is not None:
-            ...     calls = options[options['option_type'] == 'call']
-            ...     puts = options[options['option_type'] == 'put']
-        """
-        try:
-            # Try database first (primary source as of Issue #373)
-            query = """
-                SELECT
-                    strike, option_type, expiration,
-                    bid, ask, last, volume, open_interest,
-                    implied_volatility, delta, gamma, theta, vega, rho,
-                    contract_symbol, underlying_price
-                FROM raw_options_chain
-                WHERE symbol = ? AND trading_date = ?
-            """
-            params = [symbol.upper(), trading_date]
-
-            if source:
-                query += " AND source = ?"
-                params.append(source)
-
-            query += " ORDER BY strike, option_type"
-
-            with sqlite3.connect(self.db_path) as conn:
-                df = pd.read_sql_query(query, conn, params=params)
-
-            if not df.empty:
-                self.logger.debug(
-                    f"Database hit: {len(df)} raw options for {symbol} {trading_date}"
-                )
-                return df
-
-            # Fallback to file cache (deprecated, for backward compatibility)
-            self.logger.debug(
-                f"Database miss for {symbol} {trading_date}, trying file cache fallback"
-            )
-            file_cache_df = self._get_raw_options_from_file_cache(symbol, trading_date)
-
-            if file_cache_df is not None:
-                self.logger.warning(
-                    f"Using deprecated file cache for {symbol} {trading_date}. "
-                    f"Consider migrating to database with store_raw_options()"
-                )
-                return file_cache_df
-
-            self.logger.debug(f"No raw options found for {symbol} {trading_date}")
-            return None
-
-        except Exception as e:
-            self.logger.error(f"Error retrieving raw options for {symbol} {trading_date}: {e}")
-            return None
-
-    def _get_raw_options_from_file_cache(
-        self, symbol: str, trading_date: str
-    ) -> Optional[pd.DataFrame]:
-        """
-        Fallback to file cache for raw options data (deprecated).
-
-        This provides backward compatibility during migration to database storage.
-
-        Args:
-            symbol: Stock symbol
-            trading_date: Trading date in YYYY-MM-DD format
-
-        Returns:
-            DataFrame from file cache, or None if not found
-        """
-        try:
-            # Try common file cache locations
-            file_cache_paths = [
-                Path(f".cache/options/{symbol.upper()}/{trading_date}.pickle"),
-                Path(f".cache/options/{symbol.upper()}/{trading_date}.json"),
-            ]
-
-            for cache_path in file_cache_paths:
-                if cache_path.exists():
-                    if cache_path.suffix == ".pickle":
-                        df = pd.read_pickle(cache_path)  # nosec B301
-                    elif cache_path.suffix == ".json":
-                        df = pd.read_json(cache_path)
-                    else:
-                        continue
-
-                    if not df.empty:
-                        self.logger.debug(f"File cache hit: {cache_path}")
-                        return df
-
-            return None
-
-        except Exception as e:
-            self.logger.warning(f"Error reading file cache: {e}")
-            return None
-
-    def delete_raw_options(self, symbol: str, trading_date: str = None, source: str = None) -> int:
-        """
-        Delete raw options data.
-
-        Args:
-            symbol: Stock symbol
-            trading_date: Trading date (None = all dates)
-            source: Data source filter (None = all sources)
-
-        Returns:
-            Number of rows deleted
-        """
-        try:
-            query = "DELETE FROM raw_options_chain WHERE symbol = ?"
-            params = [symbol.upper()]
-
-            if trading_date:
-                query += " AND trading_date = ?"
-                params.append(trading_date)
-
-            if source:
-                query += " AND source = ?"
-                params.append(source)
-
-            with sqlite3.connect(self.db_path) as conn:
-                cursor = conn.execute(query, params)
-                deleted = cursor.rowcount
-                conn.commit()
-
-            self.logger.debug(f"Deleted {deleted} raw options for {symbol}")
-            return deleted
-
-        except Exception as e:
-            self.logger.error(f"Error deleting raw options for {symbol}: {e}")
-            return 0
-
-    def get_options_stats(self) -> Dict[str, Any]:
-        """
-        Get raw options cache statistics.
-
-        Returns:
-            Dictionary with options metrics:
-            - total_contracts: Total number of option contracts
-            - unique_symbols: Number of unique underlying symbols
-            - trading_dates: Number of unique trading dates
-            - sources: Breakdown by data source
-            - date_range: Min and max trading dates
-        """
-        try:
-            with sqlite3.connect(self.db_path) as conn:
-                # Total contracts
-                total = conn.execute("SELECT COUNT(*) FROM raw_options_chain").fetchone()[0]
-
-                # Unique symbols
-                unique_symbols = conn.execute(
-                    "SELECT COUNT(DISTINCT symbol) FROM raw_options_chain"
-                ).fetchone()[0]
-
-                # Unique trading dates
-                unique_dates = conn.execute(
-                    "SELECT COUNT(DISTINCT trading_date) FROM raw_options_chain"
-                ).fetchone()[0]
-
-                # Breakdown by source
-                sources = {}
-                for row in conn.execute(
-                    "SELECT source, COUNT(*) as count FROM raw_options_chain GROUP BY source"
-                ):
-                    sources[row[0]] = row[1]
-
-                # Date range
-                date_range = conn.execute(
-                    "SELECT MIN(trading_date), MAX(trading_date) FROM raw_options_chain"
-                ).fetchone()
-
-            return {
-                "total_contracts": total,
-                "unique_symbols": unique_symbols,
-                "trading_dates": unique_dates,
-                "sources": sources,
-                "date_range": (
-                    {"min_date": date_range[0], "max_date": date_range[1]}
-                    if date_range[0]
-                    else None
-                ),
-            }
-
-        except Exception as e:
-            self.logger.error(f"Error getting options stats: {e}")
-            return {}
-
-    # === TRADE HISTORY METHODS (Issue #373 extension - hybrid trade storage) ===
-
-    def archive_trade(self, trade_data: Dict[str, Any]) -> bool:
-        """
-        Archive completed trade to database for analytics.
-
-        This implements the hybrid storage approach:
-        - Active trades in JSON (fast, simple)
-        - Completed trades in database (analytics, TradingView-style charts)
-
-        Args:
-            trade_data: Dictionary with trade details including:
-                - trade_id (required): Unique trade identifier
-                - symbol (required): Stock symbol
-                - entry_date (required): Entry datetime ISO format
-                - entry_price (required): Entry price
-                - quantity (required): Number of shares
-                - exit_date: Exit datetime ISO format
-                - exit_price: Exit price
-                - exit_reason: Reason for exit
-                - initial_stop_loss: Initial stop loss price
-                - initial_take_profit: Initial take profit price
-                - realized_pnl: Realized profit/loss
-                - realized_pnl_pct: Realized P&L percentage
-                - strategy_name: Strategy used (e.g., "VoterAgent")
-                - signal_strength: Signal strength (BULLISH/BEARISH/NEUTRAL)
-                - signal_confidence: Confidence level 0.0-1.0
-                - broker_account: Broker account identifier
-                - notes: Additional metadata (dict, will be JSON serialized)
-
-        Returns:
-            True if archived successfully, False otherwise
-
-        Example:
-            >>> trade_data = {
-            ...     'trade_id': 'TQQQ_2024-01-15T10:30:00',
-            ...     'symbol': 'TQQQ',
-            ...     'entry_date': '2024-01-15T10:30:00',
-            ...     'entry_price': 50.25,
-            ...     'quantity': 100,
-            ...     'exit_date': '2024-01-15T15:45:00',
-            ...     'exit_price': 52.10,
-            ...     'exit_reason': 'take_profit',
-            ...     'realized_pnl': 185.0,
-            ...     'realized_pnl_pct': 3.68,
-            ...     'strategy_name': 'VoterAgent'
-            ... }
-            >>> cache.archive_trade(trade_data)
-        """
-        try:
-            # Validate required fields
-            required_fields = ["trade_id", "symbol", "entry_date", "entry_price", "quantity"]
-            for field in required_fields:
-                if field not in trade_data:
-                    self.logger.error(f"Missing required field: {field}")
-                    return False
-
-            # Serialize notes to JSON if it's a dict
-            notes = trade_data.get("notes")
-            if isinstance(notes, dict):
-                notes = json.dumps(notes)
-
-            with self._write_lock:
-                with sqlite3.connect(self.db_path) as conn:
-                    conn.execute(
-                        """
-                        INSERT INTO trade_history (
-                            trade_id, symbol, asset_type,
-                            entry_date, entry_price, entry_order_id, quantity,
-                            exit_date, exit_price, exit_order_id, exit_reason,
-                            initial_stop_loss, initial_take_profit,
-                            final_stop_loss, final_take_profit, stop_adjustments,
-                            realized_pnl, realized_pnl_pct,
-                            max_profit_pct, max_drawdown_pct, holding_period_hours,
-                            strategy_name, signal_strength, signal_confidence,
-                            entry_slippage_pct, exit_slippage_pct, commission_paid,
-                            broker_account, notes
-                        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                    """,
-                        (
-                            trade_data["trade_id"],
-                            trade_data["symbol"].upper(),
-                            trade_data.get("asset_type", "stock"),
-                            trade_data["entry_date"],
-                            trade_data["entry_price"],
-                            trade_data.get("entry_order_id"),
-                            trade_data["quantity"],
-                            trade_data.get("exit_date"),
-                            trade_data.get("exit_price"),
-                            trade_data.get("exit_order_id"),
-                            trade_data.get("exit_reason"),
-                            trade_data.get("initial_stop_loss"),
-                            trade_data.get("initial_take_profit"),
-                            trade_data.get("final_stop_loss"),
-                            trade_data.get("final_take_profit"),
-                            trade_data.get("stop_adjustments", 0),
-                            trade_data.get("realized_pnl"),
-                            trade_data.get("realized_pnl_pct"),
-                            trade_data.get("max_profit_pct"),
-                            trade_data.get("max_drawdown_pct"),
-                            trade_data.get("holding_period_hours"),
-                            trade_data.get("strategy_name"),
-                            trade_data.get("signal_strength"),
-                            trade_data.get("signal_confidence"),
-                            trade_data.get("entry_slippage_pct"),
-                            trade_data.get("exit_slippage_pct"),
-                            trade_data.get("commission_paid"),
-                            trade_data.get("broker_account", "alpaca_paper"),
-                            notes,
-                        ),
-                    )
-                    conn.commit()
-
-            self.logger.info(
-                f"Archived trade {trade_data['trade_id']} to database: "
-                f"{trade_data['symbol']} P&L: {trade_data.get('realized_pnl', 'N/A')}"
-            )
-            return True
-
-        except sqlite3.IntegrityError as e:
-            self.logger.warning(f"Trade {trade_data.get('trade_id')} already archived: {e}")
-            return False
-        except Exception as e:
-            self.logger.error(f"Failed to archive trade: {e}", exc_info=True)
-            return False
-
-    def get_trade_history(
-        self,
-        symbol: str = None,
-        start_date: str = None,
-        end_date: str = None,
-        strategy: str = None,
-        broker_account: str = None,
-        limit: int = None,
-    ) -> pd.DataFrame:
-        """
-        Query trade history for analytics and TradingView-style visualizations.
-
-        Args:
-            symbol: Filter by symbol (None = all symbols)
-            start_date: Filter by entry date >= start_date (YYYY-MM-DD)
-            end_date: Filter by entry date <= end_date (YYYY-MM-DD)
-            strategy: Filter by strategy name
-            broker_account: Filter by broker account
-            limit: Limit number of results (None = all)
-
-        Returns:
-            DataFrame with trade history
-
-        Example:
-            >>> # Get all TQQQ trades from 2024
-            >>> trades = cache.get_trade_history(symbol="TQQQ", start_date="2024-01-01")
-            >>> print(f"Total trades: {len(trades)}")
-            >>> print(f"Win rate: {(trades['realized_pnl'] > 0).sum() / len(trades) * 100:.1f}%")
-            >>>
-            >>> # Get VoterAgent performance
-            >>> voter_trades = cache.get_trade_history(strategy="VoterAgent")
-            >>> print(f"Total P&L: ${voter_trades['realized_pnl'].sum():.2f}")
-        """
-        try:
-            query = "SELECT * FROM trade_history WHERE 1=1"
-            params = []
-
-            if symbol:
-                query += " AND symbol = ?"
-                params.append(symbol.upper())
-            if start_date:
-                query += " AND entry_date >= ?"
-                params.append(start_date)
-            if end_date:
-                query += " AND entry_date <= ?"
-                params.append(end_date)
-            if strategy:
-                query += " AND strategy_name = ?"
-                params.append(strategy)
-            if broker_account:
-                query += " AND broker_account = ?"
-                params.append(broker_account)
-
-            query += " ORDER BY entry_date DESC"
-
-            if limit:
-                query += f" LIMIT {int(limit)}"
-
-            with sqlite3.connect(self.db_path) as conn:
-                df = pd.read_sql_query(query, conn, params=params)
-
-            self.logger.debug(f"Retrieved {len(df)} trades from history")
-            return df
-
-        except Exception as e:
-            self.logger.error(f"Error querying trade history: {e}")
-            return pd.DataFrame()
-
-    def get_trade_stats(self, symbol: str = None, strategy: str = None) -> Dict[str, Any]:
-        """
-        Get aggregated trading statistics.
-
-        Args:
-            symbol: Filter by symbol (None = all symbols)
-            strategy: Filter by strategy (None = all strategies)
-
-        Returns:
-            Dictionary with trading statistics:
-            - total_trades: Total number of trades
-            - winning_trades: Number of profitable trades
-            - losing_trades: Number of losing trades
-            - win_rate_pct: Win rate percentage
-            - total_pnl: Total realized P&L
-            - avg_pnl: Average P&L per trade
-            - avg_win: Average winning trade
-            - avg_loss: Average losing trade
-            - profit_factor: Gross profit / gross loss
-            - avg_holding_hours: Average holding period
-
-        Example:
-            >>> stats = cache.get_trade_stats(strategy="VoterAgent")
-            >>> print(f"Win Rate: {stats['win_rate_pct']:.1f}%")
-            >>> print(f"Profit Factor: {stats['profit_factor']:.2f}")
-        """
-        try:
-            where_clauses = []
-            params = []
-
-            if symbol:
-                where_clauses.append("symbol = ?")
-                params.append(symbol.upper())
-            if strategy:
-                where_clauses.append("strategy_name = ?")
-                params.append(strategy)
-
-            where_sql = " AND ".join(where_clauses) if where_clauses else "1=1"
-
-            with sqlite3.connect(self.db_path) as conn:
-                # Basic counts
-                stats_query = f"""
-                    SELECT
-                        COUNT(*) as total_trades,
-                        SUM(CASE WHEN realized_pnl > 0 THEN 1 ELSE 0 END) as winning_trades,
-                        SUM(CASE WHEN realized_pnl < 0 THEN 1 ELSE 0 END) as losing_trades,
-                        SUM(realized_pnl) as total_pnl,
-                        AVG(realized_pnl) as avg_pnl,
-                        AVG(CASE WHEN realized_pnl > 0 THEN realized_pnl END) as avg_win,
-                        AVG(CASE WHEN realized_pnl < 0 THEN realized_pnl END) as avg_loss,
-                        SUM(CASE WHEN realized_pnl > 0 THEN realized_pnl ELSE 0 END) as gross_profit,
-                        SUM(CASE WHEN realized_pnl < 0 THEN ABS(realized_pnl) ELSE 0 END) as gross_loss,
-                        AVG(holding_period_hours) as avg_holding_hours
-                    FROM trade_history
-                    WHERE {where_sql}
-                """  # nosec B608
-                row = conn.execute(stats_query, params).fetchone()
-
-            total_trades = row[0] or 0
-            winning_trades = row[1] or 0
-            losing_trades = row[2] or 0
-            total_pnl = row[3] or 0.0
-            avg_pnl = row[4] or 0.0
-            avg_win = row[5] or 0.0
-            avg_loss = row[6] or 0.0
-            gross_profit = row[7] or 0.0
-            gross_loss = row[8] or 0.0
-            avg_holding_hours = row[9] or 0.0
-
-            win_rate_pct = (winning_trades / total_trades * 100) if total_trades > 0 else 0.0
-            profit_factor = (gross_profit / gross_loss) if gross_loss > 0 else 0.0
-
-            return {
-                "total_trades": total_trades,
-                "winning_trades": winning_trades,
-                "losing_trades": losing_trades,
-                "win_rate_pct": win_rate_pct,
-                "total_pnl": total_pnl,
-                "avg_pnl": avg_pnl,
-                "avg_win": avg_win,
-                "avg_loss": avg_loss,
-                "profit_factor": profit_factor,
-                "avg_holding_hours": avg_holding_hours,
-            }
-
-        except Exception as e:
-            self.logger.error(f"Error getting trade stats: {e}")
-            return {}
+    # === OPTIONS AND TRADE METHODS: See OptionsCacheMixin and TradeCacheMixin (Issue #510) ===
 
     def _migrate_add_timeframe(self, conn: sqlite3.Connection):
         """
